refactor(plugin-webpack): move webpack config generation code to its own file
