+++ packages/plugin/webpack/test/WebpackPlugin_spec.ts
@@ -6,55 +6,23 @@ import { tmpdir } from 'os';
 import WebpackPlugin from '../src/WebpackPlugin';
 
 describe('WebpackPlugin', () => {
+  const baseConfig = {
+    mainConfig: {},
+    renderer: {
+      config: {},
+      entryPoints: [],
+    },
+  };
+
   const webpackTestDir = path.resolve(tmpdir(), 'electron-forge-plugin-webpack-test');
 
-  describe('PRELOAD_WEBPACK_ENTRY', () => {
-    it('should assign absolute preload script path in development', () => {
-      const p = new WebpackPlugin({
-        mainConfig: {},
-        renderer: {
-          config: {},
-          entryPoints: [
-            {
-              js: 'window.js',
-              name: 'window',
-              preload: {
-                js: 'preload.js',
-              },
-            },
-          ],
-        },
-      });
-      p.init(process.platform === 'win32' ? 'C:\\baseDir' : '/baseDir');
-      const defines = p.getDefines();
-
-      if (process.platform === 'win32') {
-        expect(defines.WINDOW_PRELOAD_WEBPACK_ENTRY).to.be.eq(String.raw`'C:\\baseDir\\.webpack\\renderer\\window\\preload.js'`);
-      } else {
-        expect(defines.WINDOW_PRELOAD_WEBPACK_ENTRY).to.be.eq("'/baseDir/.webpack/renderer/window/preload.js'");
-      }
+  describe('TCP port', () => {
+    it('should fail for privileged ports', () => {
+      expect(() => new WebpackPlugin({ ...baseConfig, loggerPort: 80 })).to.throw(/privileged$/);
     });
 
-    it('should assign an expression to resolve the preload script in production', () => {
-      const p = new WebpackPlugin({
-        mainConfig: {},
-        renderer: {
-          config: {},
-          entryPoints: [
-            {
-              js: 'window.js',
-              name: 'window',
-              preload: {
-                js: 'preload.js',
-              },
-            },
-          ],
-        },
-      });
-      p.init(process.platform === 'win32' ? 'C:\\baseDir' : '/baseDir');
-      (p as any).isProd = true;
-      const defines = p.getDefines();
-      expect(defines.WINDOW_PRELOAD_WEBPACK_ENTRY).to.be.eq("require('path').resolve(__dirname, '../renderer', 'window', 'preload.js')");
+    it('should fail for too-large port numbers', () => {
+      expect(() => new WebpackPlugin({ ...baseConfig, loggerPort: 99999 })).to.throw(/not a valid TCP port/);
     });
   });
 
@@ -66,13 +34,7 @@ describe('WebpackPlugin', () => {
 
     before(async () => {
       await fs.ensureDir(packagedPath);
-      plugin = new WebpackPlugin({
-        mainConfig: {},
-        renderer: {
-          config: {},
-          entryPoints: [],
-        },
-      });
+      plugin = new WebpackPlugin(baseConfig);
       plugin.setDirectories(webpackTestDir);
     });
